Cryptofy
This application is responsible for encrypting and decrypting text. It was created with vanilla JS, Tailwind CSS and HTML as part of an Alura Challenge One challenge. The application allows the user to enter a text, and choose to encrypt or decrypt the text. The result is displayed on a card which contains the text. The application has a simple and modern design, using the Tailwind CSS utility classes to create a consistent layout system adaptable to different screen sizes.